\name{auto.pca}
\alias{auto.pca}
\title{Automatic Variable Reduction Using Principal Component Analysis}
\usage{
auto.pca(input_data)
}
\arguments{
  \item{input_data}{dataframe without ID Variables & Categorical Variables}
}
\description{
Prints the uncorrelated variables from the input dataframe
}
\examples{
auto.pca(attitude)
}
